Eleven: Add Section Headers
This implements section headers for the song/album/artist fragments for each of the sort types
including dates, numbers, durations, alphabetical etc.  The ui will need to be tweaked to look
better, but that is pending UX.  There is a bug with localization (like Chinese songs)
where we could create a valid section header but from an alphabetical it doesn't look correct.
I will look at this afer this big check-in

Change-Id: Ibfc09bb12ebf61c668f135358f65f1e7947bac79