Eleven: This adds swiping gestures to the music player in the main area
Adds a view pager with album art to the main browse area.  This also fixes a few
bugs around the queueing/playback logic without re-writing.  After the demo, we need
to sit down and analyze whether the compromised solution is sufficient or whether we
want to do a full re-write of the queueing logic.  For now, it has been patched enough
that to the user there are some minor kinks but are pretty rare and more of an annoyance
than an actual problem AFAIK

Change-Id: Ifad19e222714bbe76c4d55708d79b035bad06777