Eleven: This adds the playlist top artist and song artwork logic
Reshuffles a lot of code from ImageWorker/ImageFetcher into ImageUtils and
Separate worker tasks.  The artwork logic leverages the top songs db to
figure the top artist as well as the top songs for a playlist and builds images
based on them.  There is some basic logic to prevent this code from running all
the time

https://cyanogen.atlassian.net/browse/MUSIC-27

Change-Id: I05d832d4f0810b71ed0c0df184a372771537ccc1