Eleven: Update the lists when things change
https://cyanogen.atlassian.net/browse/MUSIC-50

There is still a race condition where you delete an artist and immediately click it
and can get it to crash.  Not quite sure how to fix it as restarting the loader doesn't
do it.  The only thing I can think of is temporarily remove it from the adapter and also
restart the loader.  Since this is less likely scenario I've separated it into its own bug
https://cyanogen.atlassian.net/browse/MUSIC-76

Change-Id: I4a1a705d311c991686b1aa5ec02ff6a1a9f9241d