Eleven: Fix some bugs around the blur drawable not loading
I found three main bugs after some painful investigation

1) The stored key was different from the task key causing the check to always
return false, meaning the blurtask was always cancelled.  If we tried to load
the same album music 3 times in a row, the blur image will always re-kick off
even though they are the same image

2) Images don't load while scrolling for perf reasons, and for the lists,
the adapters are notified on scroll stop, but the blur image/album art pager/
bottom album art aren't

3) This was the winner - so if we load an image followed by two default images
on the 2nd default image we create the transition drawable but don't set it into
the blur image because it is a duplicate.  Inside the transition drawable it takes
the drawable and sets the listener to the transition drawable, so even if though we
don't set it into the imageview, the 1st transition drawable now isn't hooked up properly
and if this happens while the application isn't in the foreground then you can come back
and the transition drawable wouldn't have fired off

https://cyanogen.atlassian.net/browse/MUSIC-173
https://cyanogen.atlassian.net/browse/MUSIC-138

Change-Id: I9f99273bfcfe328179849f2131d4b1b2c98ca088