Settings: expose missing "wifi only" setting to user
When loading images, we try to fetch them from the caches (disk and lru).
If an image is not found, we fetch it from the internet.
Prior to this, we call "ApolloUtils.isOnline()", however this method
checks for a non exposed preference "wifiOnly", which defaults to true.

This results in a requirement of the user to be on wifi in order to fetch
images and thus this function is broken on mobile data.

To work against that, expose the wifi only option to the user.

Change-Id: Ice0a09e0c28b024caf497dc1a80b6d873cfb6287