Generate all color swatches.
While we only use the vibrant colors in the UI, we also use the palette
to determine the dominant color and compare the vibrant colors against
that. If the dominant color is one of the muted swatches, by only generating
the vibrant swatches we needlessly exclude one of the vibrant colors
from being used as visualizer color.

Change-Id: I3ca1059a653ff39b49b6a38a934863ef6ad2d976